requirements.txt
pyusb==1.3.1
pillow==12.2.0
